fix(identity): make principal deletion durable
